requirements.txt
bcrypt
requests
PyYAML
psutil
python-dotenv
tqdm
watchdog
transformers
safetensors
